+++ include/morphio/shared_utils.tpp
@@ -12,23 +12,23 @@ float _somaSurface(const SomaType type, const ContainerDiameters& diameters,
 
     switch (type) {
     case SOMA_SINGLE_POINT: {
-        float radius = diameters[0] / 2;
-        return 4 * M_PI * radius * radius;
+        float radius = diameters[0] / 2.f;
+        return 4.f * static_cast<float>(M_PI) * radius * radius;
     }
 
     case SOMA_NEUROMORPHO_THREE_POINT_CYLINDERS: {
-        float radius = diameters[0] / 2;
-        return 4 * M_PI * radius * radius;
+        float radius = diameters[0] / 2.f;
+        return 4.f * static_cast<float>(M_PI) * radius * radius;
     }
     case SOMA_CYLINDERS: {
         // Surface is approximated as the sum of areas of the conical frustums
         // defined by each segments. Does not include the endcaps areas
         float surface = 0;
         for (unsigned int i = 0; i < size - 1; ++i) {
-            float r0 = diameters[i] * 0.5;
-            float r1 = diameters[i + 1] * 0.5;
+            float r0 = diameters[i] * 0.5f;
+            float r1 = diameters[i + 1] * 0.5f;
             float h2 = distance(points[i], points[i + 1]);
-            auto s = M_PI * (r0 + r1) * sqrt((r0 - r1) * (r0 - r1) + h2 * h2);
+            auto s = static_cast<float>(M_PI) * (r0 + r1) * sqrtf((r0 - r1) * (r0 - r1) + h2 * h2);
             surface += s;
         }
         return surface;
